Move native variant rules to respective plugin
